requirements.txt
python-dotenv
pymongo
mysql-connector
